tps62360-regulator: Fix pointer-to-int cast
Casting match->data to an integer directly creates an int to pointer warning.
Cast it to intptr_t then int to get rid of warning.

+++ drivers/regulator/tps62360-regulator.c
@@ -367,7 +367,7 @@ static int tps62360_probe(struct i2c_client *client,
 			dev_err(&client->dev, "Error: No device match found\n");
 			return -ENODEV;
 		}
-		chip_id = (int)match->data;
+		chip_id = (int)(intptr_t)match->data;
 		if (!pdata)
 			pdata = of_get_tps62360_platform_data(&client->dev);
 	}
